- Changed names of tables; - Changed names of entities; - Added new change sets to the db-migration.xml
