+++ src/main/kotlin/com/localization/model/Country.kt
@@ -8,13 +8,7 @@ class Country(
     @field:Id
     @field:GeneratedValue(strategy = GenerationType.IDENTITY)
     var id: Long?,
-    var name: String?,
-    @field:OneToOne
-    @field:JoinColumn(name = "language_id")
-    var language: Language?,
-    @field:OneToOne
-    @field:JoinColumn(name = "iso_id")
-    var isoCode: IsoCode?
-){
-    constructor() : this(null, null, null, null)
+    var iso: String?
+) {
+    constructor(): this(null, null)
 }
