+++ src/main/kotlin/com/localization/repository/CountryRepository.kt
@@ -1,17 +1,17 @@
 package com.localization.repository
 
-import com.localization.model.Country
+import com.localization.model.CountryLocalization
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Query
 import org.springframework.data.repository.query.Param
 
-interface CountryRepository : JpaRepository<Country, Int> {
+interface CountryRepository : JpaRepository<CountryLocalization, Int> {
     @Query(
-        value = "FROM Country c LEFT JOIN FETCH c.language LEFT JOIN FETCH c.isoCode " +
-                "WHERE c.language.language = :language AND c.isoCode.iso = :isoCode"
+        value = "FROM CountryLocalization c LEFT JOIN FETCH c.language LEFT JOIN FETCH c.country " +
+                "WHERE c.language.language = :language AND c.country.iso = :isoCode"
     )
     fun getByIsoCodeAndLanguage(
         @Param("isoCode") isoCode: String,
         @Param("language") language: String
-    ): Country?
+    ): CountryLocalization?
 }
