+++ src/main/kotlin/com/localization/repository/IsoCodeRepository.kt
@@ -1,8 +1,8 @@
 package com.localization.repository
 
-import com.localization.model.IsoCode
+import com.localization.model.Country
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface IsoCodeRepository: JpaRepository<IsoCode, Long> {
+interface IsoCodeRepository: JpaRepository<Country, Long> {
     fun existsByIso(isoCode: String): Boolean
 }
