+++ src/main/kotlin/com/localization/service/CountryService.kt
@@ -1,7 +1,7 @@
 package com.localization.service
 
-import com.localization.model.Country
+import com.localization.model.CountryLocalization
 
 interface CountryService {
-    fun getLocalizationByLanguageAndIso(isoCode: String, language: String): Country
+    fun getLocalizationByLanguageAndIso(isoCode: String, language: String): CountryLocalization
 }
