+++ src/main/kotlin/com/localization/service/mapper/CountryMapToDto.kt
@@ -1,8 +1,8 @@
 package com.localization.service.mapper
 
-import com.localization.model.Country
+import com.localization.model.CountryLocalization
 import com.localization.model.dto.CountryResponseDto
 
 interface CountryMapToDto {
-    fun mapToDto(country: Country): CountryResponseDto
+    fun mapToDto(countryLocalization: CountryLocalization): CountryResponseDto
 }
