+++ src/main/kotlin/com/localization/service/mapper/impl/CountryMapToDtoImpl.kt
@@ -1,11 +1,11 @@
 package com.localization.service.mapper.impl
 
-import com.localization.model.Country
+import com.localization.model.CountryLocalization
 import com.localization.model.dto.CountryResponseDto
 import com.localization.service.mapper.CountryMapToDto
 import org.springframework.stereotype.Component
 
 @Component
 class CountryMapToDtoImpl : CountryMapToDto {
-    override fun mapToDto(country: Country): CountryResponseDto = CountryResponseDto(country.name)
+    override fun mapToDto(countryLocalization: CountryLocalization): CountryResponseDto = CountryResponseDto(countryLocalization.name)
 }
